requirements.txt
asgiref==3.2.10
Django==3.1.1
django-tastypie==0.14.3
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
Faker==4.1.3
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
